# apis/app-api/payment/sdkwork-payment-app-api.openapi.yaml
{
  "openapi": "3.1.2",
  "info": {
    "title": "SDKWork Payment App API",
    "version": "1.0.0",
    "description": "App/client payment execution and refund contract for commerce.payment.",
    "x-sdkwork-api-authority": "sdkwork-payment-app-api",
    "x-sdkwork-sdk-family": "sdkwork-payment-app-sdk",
    "x-sdkwork-audience": "Authenticated SDKWork payment clients"
  },
  "servers": [{ "url": "http://localhost:18094", "description": "Local payment gateway" }],
...
  "security": [{ "AuthToken": [], "AccessToken": [] }],
  "paths": {
    "/app/v3/api/payments/intents": {
      "post": {
        "tags": ["commerce"], "summary": "Create a payment intent.", "operationId": "payments.intents.create",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreatePaymentIntentCommand" } } } },
        "responses": { "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PaymentIntentResponse" } } } }, "default": { "$ref": "#/components/responses/Problem" } },
        "security": [{ "AuthToken": [], "AccessToken": [] }], "x-sdkwork-auth-mode": "dual-token", "x-sdkwork-forbid-credential-headers": false, "x-sdkwork-owner": "sdkwork-payment", "x-sdkwork-api-authority": "sdkwork-payment-app-api", "x-sdkwork-domain": "commerce", "x-sdkwork-resource": "paymentIntents", "x-sdkwork-request-context": "WebRequestContext", "x-sdkwork-api-surface": "app-api", "x-sdkwork-server-request-id": true, "x-sdkwork-permission": "commerce.payments.intents.create", "x-sdkwork-idempotent": true
      }
    },
    "/app/v3/api/payments/intents/{paymentIntentId}": {
      "get": {
        "tags": ["commerce"], "summary": "Retrieve a payment intent.", "operationId": "payments.intents.retrieve",
        "parameters": [{ "$ref": "#/components/parameters/PaymentIntentId" }],
        "responses": { "200": { "description": "Success", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PaymentIntentResponse" } } } }, "default": { "$ref": "#/components/responses/Problem" } },
        "security": [{ "AuthToken": [], "AccessToken": [] }], "x-sdkwork-auth-mode": "dual-token", "x-sdkwork-forbid-credential-headers": false, "x-sdkwork-owner": "sdkwork-payment", "x-sdkwork-api-authority": "sdkwork-payment-app-api", "x-sdkwork-domain": "commerce", "x-sdkwork-resource": "paymentIntents", "x-sdkwork-request-context": "WebRequestContext", "x-sdkwork-api-surface": "app-api", "x-sdkwork-server-request-id": true, "x-sdkwork-permission": "commerce.payments.intents.read"
      }
    },
    "/app/v3/api/payments/intents/{paymentIntentId}/cancel": {
      "post": {
        "tags": ["commerce"], "summary": "Cancel a payment intent.", "operationId": "payments.intents.cancel",
        "parameters": [{ "$ref": "#/components/parameters/PaymentIntentId" }],
        "responses": { "200": { "description": "Accepted", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SdkWorkCommandResponse" } } } }, "default": { "$ref": "#/components/responses/Problem" } },
        "security": [{ "AuthToken": [], "AccessToken": [] }], "x-sdkwork-auth-mode": "dual-token", "x-sdkwork-forbid-credential-headers": false, "x-sdkwork-owner": "sdkwork-payment", "x-sdkwork-api-authority": "sdkwork-payment-app-api", "x-sdkwork-domain": "commerce", "x-sdkwork-resource": "paymentIntents", "x-sdkwork-request-context": "WebRequestContext", "x-sdkwork-api-surface": "app-api", "x-sdkwork-server-request-id": true, "x-sdkwork-permission": "commerce.payments.intents.cancel", "x-sdkwork-idempotent": true
      }
    },
    "/app/v3/api/payments/intents/{paymentIntentId}/attempts": {
      "post": {
        "tags": ["commerce"], "summary": "Create a payment attempt.", "operationId": "payments.intents.attempts.create",
        "parameters": [{ "$ref": "#/components/parameters/PaymentIntentId" }],
        "responses": { "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PaymentAttemptResponse" } } } }, "default": { "$ref": "#/components/responses/Problem" } },
        "security": [{ "AuthToken": [], "AccessToken": [] }], "x-sdkwork-auth-mode": "dual-token", "x-sdkwork-forbid-credential-headers": false, "x-sdkwork-owner": "sdkwork-payment", "x-sdkwork-api-authority": "sdkwork-payment-app-api", "x-sdkwork-domain": "commerce", "x-sdkwork-resource": "paymentAttempts", "x-sdkwork-request-context": "WebRequestContext", "x-sdkwork-api-surface": "app-api", "x-sdkwork-server-request-id": true, "x-sdkwork-permission": "commerce.payments.attempts.create", "x-sdkwork-idempotent": true
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AuthToken": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" },
      "AccessToken": { "type": "apiKey", "in": "header", "name": "Access-Token" }
    },
    "parameters": {
...
      "PaymentIntentId": { "name": "paymentIntentId", "in": "path", "required": true, "schema": { "type": "string" } },
...
    },
    "responses": {
      "Problem": { "description": "SDKWork problem detail", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ProblemDetail" } } } }
    },
    "schemas": {
      "SdkWorkApiResponse": { "type": "object", "additionalProperties": false, "required": ["code", "data", "traceId"], "properties": { "code": { "type": "integer", "format": "int32", "const": 0 }, "data": { "type": "object" }, "traceId": { "type": "string", "format": "uuid" } } },
...
      "SdkWorkCommandData": { "type": "object", "additionalProperties": false, "required": ["accepted"], "properties": { "accepted": { "type": "boolean", "const": true }, "resourceId": { "type": "string" }, "status": { "type": "string" } } },
...
      "SdkWorkCommandResponse": { "allOf": [{ "$ref": "#/components/schemas/SdkWorkApiResponse" }, { "type": "object", "required": ["data"], "properties": { "data": { "$ref": "#/components/schemas/SdkWorkCommandData" } } }] },
...
      "ProblemDetail": { "type": "object", "additionalProperties": true, "required": ["type", "title", "status", "code", "traceId"], "properties": { "type": { "type": "string", "format": "uri-reference" }, "title": { "type": "string" }, "status": { "type": "integer", "minimum": 400, "maximum": 599 }, "detail": { "type": "string" }, "code": { "type": "integer", "format": "int32", "minimum": 40001, "maximum": 79999 }, "traceId": { "type": "string", "format": "uuid" } } },
      "CreatePaymentIntentCommand": { "type": "object", "additionalProperties": false, "required": ["orderId"], "properties": { "orderId": { "type": "string" }, "paymentMethod": { "type": "string", "default": "wechat_native" } } },
...
      "PaymentIntent": { "type": "object", "additionalProperties": false, "required": ["paymentIntentId", "orderId", "paymentIntentNo", "paymentMethod", "providerCode", "amount", "currencyCode", "status"], "properties": { "paymentIntentId": { "type": "string" }, "orderId": { "type": "string" }, "paymentIntentNo": { "type": "string" }, "paymentMethod": { "type": "string" }, "providerCode": { "type": "string" }, "amount": { "type": "string" }, "currencyCode": { "type": "string" }, "status": { "type": "string" } } },
      "PaymentAttempt": { "type": "object", "additionalProperties": false, "required": ["attemptId", "orderId", "paymentMethod", "amount", "status"], "properties": { "attemptId": { "type": "string" }, "paymentIntentId": { "type": "string" }, "orderId": { "type": "string" }, "outTradeNo": { "type": "string" }, "paymentMethod": { "type": "string" }, "providerCode": { "type": "string" }, "amount": { "type": "string" }, "status": { "type": "string" }, "paymentParams": { "type": "object", "additionalProperties": { "type": "string" } } } },
...
      "PaymentIntentResponse": { "allOf": [{ "$ref": "#/components/schemas/SdkWorkApiResponse" }, { "type": "object", "required": ["data"], "properties": { "data": { "type": "object", "required": ["item"], "properties": { "item": { "$ref": "#/components/schemas/PaymentIntent" } } } } }] },
      "PaymentAttemptResponse": { "allOf": [{ "$ref": "#/components/schemas/SdkWorkApiResponse" }, { "type": "object", "required": ["data"], "properties": { "data": { "type": "object", "required": ["item"], "properties": { "item": { "$ref": "#/components/schemas/PaymentAttempt" } } } } }] },
...
    }
  }
}